Add suspend and enable ednpoints

+++ components/interop/interop-rest-service-impl/src/main/kotlin/net/corda/interop/rest/impl/v1/InteropRestResourceImpl.kt
@@ -228,7 +228,8 @@ internal class InteropRestResourceImpl @Activate constructor(
                 facadeIds = facadeIds(),
                 applicationName = createInteropIdentityRestRequest.applicationName,
                 endpointUrl = endpointUrl,
-                endpointProtocol = endpointProtocol
+                endpointProtocol = endpointProtocol,
+                enabled = true
             )
         )
 
@@ -244,6 +245,7 @@ internal class InteropRestResourceImpl @Activate constructor(
                     applicationName = MemberX500Name.parse(member.x500Name).organization,
                     endpointUrl = member.endpointUrl,
                     endpointProtocol = member.endpointProtocol,
+                    enabled = true
                 )
             )
         }
@@ -253,6 +255,46 @@ internal class InteropRestResourceImpl @Activate constructor(
         return CreateInteropIdentityRest.Response(
             Utils.computeShortHash(ownedInteropIdentityX500, interopGroupId).toString()
         )
+    }
+
+    private fun updateInteropIdentityEnablement(
+        holdingIdentityShortHash: String,
+        interopIdentityShortHash: String,
+        newState: Boolean
+    ) {
+        val validHoldingIdentityShortHash = validateShortHash(holdingIdentityShortHash)
+        val validInteropIdentityShortHash = validateShortHash(interopIdentityShortHash)
+
+        val vNodeInfo = getAndValidateVirtualNodeInfoByShortHash(validHoldingIdentityShortHash)
+        val vNodeShortHash = vNodeInfo.getVNodeShortHash()
+
+        val registryView = interopIdentityRegistryService.getVirtualNodeRegistryView(vNodeShortHash)
+
+        val identityToDisable = registryView.getIdentityWithShortHash(validInteropIdentityShortHash) ?:
+        throw InvalidInputDataException(
+            "No interop identity with short hash '$validInteropIdentityShortHash' found for holding " +
+                    "identity '$validHoldingIdentityShortHash'."
+        )
+
+        if (identityToDisable.enabled != newState) {
+            interopIdentityWriteService.updateInteropIdentityEnablement(vNodeShortHash, identityToDisable, newState)
+        }
+    }
+
+    override fun suspendInteropIdentity(
+        holdingIdentityShortHash: String,
+        interopIdentityShortHash: String
+    ): ResponseEntity<String> {
+        updateInteropIdentityEnablement(holdingIdentityShortHash, interopIdentityShortHash, false)
+        return ResponseEntity(ResponseCode.OK, "OK")
+    }
+
+    override fun enableInteropIdentity(
+        holdingIdentityShortHash: String,
+        interopIdentityShortHash: String
+    ): ResponseEntity<String> {
+        updateInteropIdentityEnablement(holdingIdentityShortHash, interopIdentityShortHash, true)
+        return ResponseEntity(ResponseCode.OK, "OK")
     }
 
     override fun deleteInteropIdentity(
@@ -271,6 +313,14 @@ internal class InteropRestResourceImpl @Activate constructor(
                 "No interop identity with short hash '$validInteropIdentityShortHash' found for holding " +
                         "identity '$validHoldingIdentityShortHash'."
             )
+
+        if (identityToRemove.enabled) {
+            throw InvalidInputDataException(
+                "Interop identity '$interopIdentityShortHash' must be disabled prior to deletion. " +
+                "Note: Deleting an interop identity will disrupt any active flow sessions which are using that identity. " +
+                "Ensure that the identity is not in use attempting to delete it."
+            )
+        }
 
         interopIdentityWriteService.removeInteropIdentity(vNodeShortHash, identityToRemove)
 
@@ -297,7 +347,8 @@ internal class InteropRestResourceImpl @Activate constructor(
                 interopIdentity.facadeIds,
                 MemberX500Name.parse(interopIdentity.x500Name).organization,
                 interopIdentity.endpointUrl,
-                interopIdentity.endpointProtocol
+                interopIdentity.endpointProtocol,
+                interopIdentity.enabled
             )
         }.toList()
     }
@@ -404,7 +455,8 @@ internal class InteropRestResourceImpl @Activate constructor(
                     owningVirtualNodeShortHash = ShortHash.of(member.owningIdentityShortHash),
                     applicationName = MemberX500Name.parse(member.x500Name).organization,
                     endpointUrl = member.endpointUrl,
-                    endpointProtocol = member.endpointProtocol
+                    endpointProtocol = member.endpointProtocol,
+                    enabled = true
                 )
             )
         }
